final review: fix major findings from the 4-agent review
All 4 reviewers returned ship_with_minors (no blockers). Fixed the major
findings; clippy now clean, build green, 100 tests pass (default and
--features local-api).

Architecture:
- models.rs: bound the first-run model download (connect_timeout 30s +
  30min overall cap) so a dead/stalled socket can no longer wedge the
  single pipeline worker forever.

Security:
- llm: send the Gemini API key in the x-goog-api-key header instead of the
  URL query string, so it can not leak into reqwest error/log output.
- tauri.conf.json: set a Content-Security-Policy (was null) restricting to
  self + IPC; blocks remote script/img/connect (defense-in-depth for the
  innerHTML dashboard). NOTE: needs runtime verification on each OS.

Build/release:
- tauri.conf.json: bundle.macOS.entitlements ../entitlements.plist ->
  entitlements.plist (Tauri resolves relative to src-tauri/; the old path
  pointed outside the project and would break macOS signing).
- .deb depends: add libayatana-appindicator3-1 (the system tray needs it).

Data safety / parity:
- db/schema.rs: only rename history.json -> .json.migrated when rows were
  ACTUALLY imported (import now returns a bool). Previously a user DB that
  already had rows would get its history.json renamed without import.
- overlay.rs: apply config.orb_position (5 presets + custom {x,y}) instead
  of hardcoding bottom-center, closing the orb-position parity gap.

Quality:
- clippy: clean (redundant i64 cast removed; intentional index loop and
  test/API-only items annotated).

Deferred as documented minors: overlay window ops run on the worker thread
(Tauri v2 proxies these; flagged for runtime verification), history stores
raw transcription (intended cli.py parity), get_statistics mode arg
(forward-compat; frontend requests all), local_api body-size cap (loopback
+ off by default).

+++ src-tauri/src/pipeline.rs
@@ -159,6 +159,9 @@ struct Worker {
 
 /// The currently-active recording session metadata.
 struct Active {
+    /// Mode the recording was started in. Stop is driven by the HotkeyEvent's
+    /// own mode, so this is kept as session metadata / for debugging.
+    #[allow(dead_code)]
     mode: Mode,
     /// The config snapshot captured at Start, so Stop processes with exactly the
     /// settings that were in effect when recording began (hot-apply per-utterance).
@@ -507,7 +510,7 @@ fn iso_now() -> String {
 fn civil_from_days(z: i64) -> (i64, u32, u32) {
     let z = z + 719_468;
     let era = if z >= 0 { z } else { z - 146_096 } / 146_097;
-    let doe = (z - era * 146_097) as i64; // [0, 146096]
+    let doe = z - era * 146_097; // [0, 146096]
     let yoe = (doe - doe / 1460 + doe / 36_524 - doe / 146_096) / 365; // [0, 399]
     let y = yoe + era * 400;
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100); // [0, 365]
